+++ src/faas-vscode-plugin/CHANGELOG.md
@@ -33,3 +33,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - Incorporated updated modal dialogue for simplified FN-CLI installation process.
+
+## 1.0.4 - 2024-02-05
+
+### Fixed
+
+- Added support to CREATE, EDIT and DELETE applications within Functions plugin instead of redirecting to console.